requirements.txt
cs50
Flask
Flask-Session
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
contextlib2==21.6.0
idna==3.3
lxml==4.7.1
numpy==1.22.1
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-random-user-agent==2022.1.23
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8